memory-forensics-toolkit/requirements.txt
volatility3
yara-python
rich
fpdf
